Move middlewares to own folder
